% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scht.R
\name{.perform_census_norm}
\alias{.perform_census_norm}
\title{Perform census normalisation}
\usage{
.perform_census_norm(gene_mat, transcript_mat)
}
\arguments{
\item{gene_mat}{Log-transformed gene expression matrix}

\item{transcript_mat}{Log-transformed transcript expression matrix}
}
\value{
A list containing:
\itemize{
\item gene_mat_census: Census-normalized gene expression matrix
\item transcript_mat_census: Census-normalized transcript expression matrix
}
}
\description{
Performs census normalization by adjusting expression values based on
the number of detected genes in each cell. This helps to account for
differences in cellular RNA content.
}
\keyword{Further}
\keyword{adjusts}
\keyword{based}
\keyword{by}
\keyword{cell.}
\keyword{coverage}
\keyword{detected}
\keyword{differences}
\keyword{each}
\keyword{for}
\keyword{genes}
\keyword{in}
\keyword{internal}
\keyword{number}
\keyword{of}
\keyword{on}
\keyword{re-scaling}
\keyword{the}
