% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scht.R
\name{.perform_log_transform}
\alias{.perform_log_transform}
\title{Perform log2 transformation}
\usage{
.perform_log_transform(gene_counts_norm, transcript_counts_norm)
}
\arguments{
\item{gene_counts_norm}{CPM-normalized gene counts matrix}

\item{transcript_counts_norm}{CPM-normalized transcript counts matrix}
}
\value{
A list containing:
\itemize{
\item gene_mat: Log2-transformed gene expression matrix
\item transcript_mat: Log2-transformed transcript expression matrix
}
}
\description{
Applies log2(x + 1) transformation to normalized count data to reduce
the effect of outliers and make the data distribution more symmetric.
}
\keyword{internal}
